cloud-function/requirements.txt
# Runtime: Python 3.9
functions-framework==3.*
google-cloud-storage==2.*
beautifulsoup4==4.*
requests==2.*
